requirements.txt
PyGithub==1.55
python-jenkins==1.7.0
requests==2.28.1
aiohttp==3.9.1
asyncio==3.4.3
llama-cpp-python>=0.3.0
python-dotenv>=1.0.0